refactor(app): group remaining shared/ files into api/forms/ui subfolders
Reorganizes what's left of the flat shared/ folder into cohesive
subfolders instead of ~15 files side by side:
- shared/api/ (api-responses.ts, used by 30+ API route handlers)
- shared/forms/ (integer-input.tsx, scientific-input.tsx,
  range-input-form.tsx, zod-resolver.ts)
- shared/ui/ (forbidden.tsx, unauthorized.tsx, not-found.tsx,
  error-dialog.tsx)

Genuinely standalone cross-cutting utilities (utils.ts, lxcatid.ts,
annotate-mixture.ts, download.ts, swr-cache-map.ts, latex.tsx) remain
at shared/ root. Updates all import paths accordingly. No behavioral
changes.
