+++ app/src/app/api/middleware/zod.ts
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-import { badRequestResponse } from "@/shared/api-responses";
+import { badRequestResponse } from "@/shared/api/api-responses";
 import { NextRequest } from "next/server";
 import { err, ok } from "true-myth/result";
 import { z } from "zod";
