+++ app/src/app/api/route-builder.ts
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-import { badRequestResponse } from "@/shared/api-responses";
+import { badRequestResponse } from "@/shared/api/api-responses";
 import { NextRequest, NextResponse } from "next/server";
 import { Result } from "true-myth";
 import { err, ok } from "true-myth/result";
