+++ app/src/app/api/scat-cs/[id]/route.ts
@@ -6,7 +6,7 @@ import { db } from "@lxcat/database";
 import {
   notFoundResponse,
   okJsonResponse,
-} from "../../../../shared/api-responses";
+} from "@/shared/api/api-responses";
 import {
   hasDeveloperOrDownloadRole,
   hasSessionOrAPIToken,
