+++ app/src/app/api/scat-css/LTPMixture/route.ts
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-import { okJsonResponse } from "@/shared/api-responses";
+import { okJsonResponse } from "@/shared/api/api-responses";
 import { LTPMixtureJSONSchema } from "@lxcat/schema/json-schema";
 import { RouteBuilder } from "../../route-builder";
