+++ app/src/app/api/states/in-reaction/route.ts
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-import { okJsonResponse } from "@/shared/api-responses";
+import { okJsonResponse } from "@/shared/api/api-responses";
 import { db } from "@lxcat/database";
 import { zodMiddleware } from "../../middleware/zod";
 import { RouteBuilder } from "../../route-builder";
