+++ app/src/app/author/set/[id]/edit/edit-form.tsx
@@ -5,7 +5,7 @@
 "use client";
 
 import { reference2bibliography } from "@/citation/cite";
-import { zodResolver } from "@/shared/zod-resolver";
+import { zodResolver } from "@/shared/forms/zod-resolver";
 import { KeyedOrganization } from "@lxcat/database/auth";
 import { EditedLTPDocument, Reference } from "@lxcat/schema";
 import { AnySpeciesSerializable } from "@lxcat/schema/species";
