+++ app/src/app/author/set/[id]/editraw/page.tsx
@@ -4,9 +4,9 @@
 
 import { options } from "@/auth/options";
 import { userIsAuthor } from "@/auth/page-guards";
-import { Forbidden } from "@/shared/forbidden";
-import { NotFound } from "@/shared/not-found";
-import { Unauthorized } from "@/shared/unauthorized";
+import { Forbidden } from "@/shared/ui/forbidden";
+import { NotFound } from "@/shared/ui/not-found";
+import { Unauthorized } from "@/shared/ui/unauthorized";
 import { db } from "@lxcat/database";
 import { intoEditable } from "@lxcat/schema/process";
 import { getServerSession } from "next-auth/next";
