+++ app/src/app/author/set/add/page.tsx
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 import { options } from "@/auth/options";
-import { Unauthorized } from "@/shared/unauthorized";
+import { Unauthorized } from "@/shared/ui/unauthorized";
 import { db } from "@lxcat/database";
 import { getServerSession } from "next-auth";
 import { EditForm } from "../[id]/edit/edit-form";
