+++ app/src/shared/forms/scientific-input.tsx
