+++ app/src/shared/ui/error-dialog.tsx
